backend-patient/src/main/java/com/medilabo/patient/controller/PatientController.java
package com.medilabo.patient.controller;

import com.medilabo.patient.model.Patient;
import com.medilabo.patient.service.PatientService;
import jakarta.validation.Valid;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.data.domain.Page;
import org.springframework.data.domain.PageRequest;
import org.springframework.data.domain.Pageable;
import org.springframework.data.domain.Sort;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

@RestController
@RequestMapping("/patients")
public class PatientController {
    @Autowired
    private PatientService patientService;

    @GetMapping
    public ResponseEntity<Page<Patient>> getPatients(@RequestParam(value = "page", defaultValue = "0") int page, @RequestParam(value = "size", defaultValue = "10") int size, @RequestParam(value = "sort", defaultValue = "id,asc") String[] sort) {
        Sort sortOrder;

        if (sort.length == 2) {
            try {
                Sort.Direction direction = Sort.Direction.fromString(sort[1]);
                sortOrder = Sort.by(new Sort.Order(direction, sort[0]));
            } catch (IllegalArgumentException e) {
                sortOrder = Sort.by("id");
            }
        } else {
            sortOrder = Sort.by("id");
        }

        Pageable pageable = PageRequest.of(page, size, sortOrder);
        Page<Patient> patientPage = patientService.getAllPatients(pageable);

        return ResponseEntity.ok(patientPage);
    }

    @GetMapping("/{id}")
    public ResponseEntity<Patient> getPatientById(@PathVariable("id") Long id) {
        return ResponseEntity.ok(patientService.getPatientById(id));
    }

    @PostMapping
    public ResponseEntity<Patient> createPatient(@Valid @RequestBody Patient patient) {
        return ResponseEntity.status(HttpStatus.CREATED).body(patientService.createPatient(patient));
    }

    @PutMapping("/{id}")
    public ResponseEntity<Patient> updatePatient(@PathVariable("id") Long id, @Valid @RequestBody Patient patient) {
        return ResponseEntity.ok(patientService.updatePatient(id, patient));
    }
}
